livekit-backend/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[all,elevenlabs,sarvam,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.5.0",
    "bcrypt>=4.1.2",
]
